Extend response with more details for answer

+++ src/main/java/io/hepp/cov2words/domain/repository/TimestampRepository.java
@@ -18,7 +18,7 @@ import java.util.UUID;
 @Repository
 public interface TimestampRepository extends JpaRepository<TimestampEntity, UUID> {
     Page<TimestampEntity> findAllByDateInvalidatedIsNullAndStatusLessThanAndDateModifiedLessThan(
-            int status,
+            long status,
             DateTime dateModified,
             Pageable pageable
     );
